requirements.txt
streamlit
pandas
matplotlib
seaborn
google-cloud-storage
